# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "@aasumitro",
            "url": "https://aasumitro.id/",
            "email": "hello@aasumitro.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/aasumitro/posbe/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/v1/floors": {
            "get": {
                "description": "Get Floors List.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Floors"
                ],
                "summary": "Floor List",
                "responses": {
                    "200": {
                        "description": "OK RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Floor"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new Floor.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Floors"
                ],
                "summary": "Store Floor Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Floor"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
        "/v1/floors/{id}": {
            "put": {
                "description": "Update Floor Data by ID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Floors"
                ],
                "summary": "Update Floor Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "floor id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Floor"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Floor Data by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Floors"
                ],
                "summary": "Delete Floor Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "floor id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT RESPOND"
                    },
                    "400": {
                        "description": "BAD REQUEST RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
        "/v1/floors/{join}": {
            "get": {
                "description": "Get Floors List With Join.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Floors"
                ],
                "summary": "Floor List With Join",
                "parameters": [
                    {
                        "enum": [
                            "rooms",
                            "tables"
                        ],
                        "type": "string",
                        "description": "join with data, available join rooms, tables",
                        "name": "join",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Floor"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Floor": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Room"
                    }
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Table"
                    }
                },
                "total_rooms": {
                    "type": "integer"
                },
                "total_tables": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                }
            }
        },
...
        "domain.Room": {
            "type": "object",
            "required": [
                "capacity",
                "floor_id",
                "h_size",
                "name",
                "price",
                "w_size",
                "x_pos",
                "y_pos"
            ],
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "floor_id": {
                    "type": "integer"
                },
                "h_size": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "w_size": {
                    "type": "number"
                },
                "x_pos": {
                    "type": "number"
                },
                "y_pos": {
                    "type": "number"
                }
            }
        },
...
        "domain.Table": {
            "type": "object",
            "required": [
                "capacity",
                "floor_id",
                "h_size",
                "name",
                "w_size",
                "x_pos",
                "y_pos"
            ],
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "created_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "floor_id": {
                    "type": "integer"
                },
                "h_size": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/sql.NullInt64"
                },
                "w_size": {
                    "type": "number"
                },
                "x_pos": {
                    "type": "number"
                },
                "y_pos": {
                    "type": "number"
                }
            }
        },
...
        "sql.NullInt64": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "utils.ErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.ValidationErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}